Ai-Expense-Tracker-v2-main/ET V2/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
alembic==1.16.2
pydantic==2.9.2
email-validator==2.2.0
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
easyocr==1.7.1
Pillow==10.4.0
python-dotenv==1.0.1
opencv-python-headless==4.10.0.84

